Split extra behavior association tests
